refactor(interface-cli): extract cmd_webui + REPL helpers
Pull two more clusters out of `main.rs`:

- `cmd_webui.rs` — thin shim that forwards `assistant webui serve`
  to `assistant_web_ui::run_from_iter` (~15 lines)
- `repl_helpers.rs` — `deliver_attachments`, `start_token_printer`,
  `print_help` used by the interactive REPL loop (~90 lines)

main.rs shrinks 1606 → 1509 lines (-97). cargo test -p assistant-cli
reports the same 59 lib tests passing; make lint is clean.

main.rs now holds: the clap argument parsing tree, top-level imports,
the cli_parse_tests module, a small `home_agent_root` helper, and the
~540-line `main()` REPL loop. The argument-parsing tree is the next
candidate slice; the REPL loop itself is the biggest remaining chunk
but tightly coupled to all the bootstrap state.
